+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/PresentBlock.java
@@ -20,4 +20,5 @@ public class PresentBlock extends AbstractPresentBlock {
         return new PresentBlockTile(pPos, pState);
     }
 
+
 }